+++ src/builders/oracle.ts
@@ -7,15 +7,9 @@ import { SUPPORT_COLLATERALS, SUPPORT_POOLS } from '../constants';
 import type { TransactionArgument } from '@mysten/sui/transactions';
 import type { SuiTxBlock as SuiKitTxBlock } from '@scallop-io/sui-kit';
 import type { ScallopBuilder } from '../models';
-import type {
-  SupportAssetCoins,
-  SupportOracleType,
-  xOracleType,
-  xOracleCategoryType,
-} from '../types';
+import type { SupportAssetCoins, SupportOracleType } from '../types';
 import { PYTH_ENDPOINTS } from 'src/constants/pyth';
 import { TransactionBlock } from '@mysten/sui.js/transactions';
-import { xOracleList } from 'src/constants';
 
 /**
  * Update the price of the oracle for multiple coin.
@@ -32,15 +26,8 @@ export const updateOracles = async (
   assetCoinNames = assetCoinNames ?? [
     ...new Set([...SUPPORT_POOLS, ...SUPPORT_COLLATERALS]),
   ];
-  // const rules: SupportOracleType[] = builder.isTestnet ? ['pyth'] : ['pyth'];
-  const flattenedRules: Set<SupportOracleType> = new Set(
-    Object.values(xOracleList).flatMap(({ primary, secondary }) => [
-      ...primary,
-      ...secondary,
-    ])
-  );
-
-  if (flattenedRules.has('pyth')) {
+  const rules: SupportOracleType[] = builder.isTestnet ? ['pyth'] : ['pyth'];
+  if (rules.includes('pyth')) {
     const pythClient = new SuiPythClient(
       builder.oldSuiClient,
       builder.address.get('core.oracles.pyth.state'),
@@ -77,12 +64,7 @@ export const updateOracles = async (
   // Remove duplicate coin names.
   const updateAssetCoinNames = [...new Set(assetCoinNames)];
   for (const assetCoinName of updateAssetCoinNames) {
-    await updateOracle(
-      builder,
-      txBlock,
-      assetCoinName,
-      xOracleList[assetCoinName]
-    );
+    await updateOracle(builder, txBlock, assetCoinName, rules);
   }
 };
 
@@ -97,7 +79,7 @@ const updateOracle = async (
   builder: ScallopBuilder,
   txBlock: SuiKitTxBlock,
   assetCoinName: SupportAssetCoins,
-  rules: xOracleType
+  rules: SupportOracleType[]
 ) => {
   const coinType = builder.utils.parseCoinType(assetCoinName);
 
@@ -142,7 +124,7 @@ const updateOracle = async (
  */
 const updatePrice = (
   txBlock: SuiKitTxBlock,
-  rules: xOracleType,
+  rules: SupportOracleType[],
   xOraclePackageId: string,
   xOracleId: TransactionArgument | string,
   pythPackageId: string,
@@ -163,42 +145,37 @@ const updatePrice = (
     xOracleId,
     coinType
   );
-  Object.entries(rules).forEach(([type, rule]: [any, SupportOracleType[]]) => {
-    if (rule.includes('pyth')) {
-      updatePythPrice(
-        type,
-        txBlock,
-        pythPackageId,
-        request,
-        pythStateId,
-        pythFeedObjectId,
-        pythRegistryId,
-        coinType
-      );
-    }
-    if (rule.includes('supra')) {
-      updateSupraPrice(
-        type,
-        txBlock,
-        supraPackageId,
-        request,
-        supraHolderId,
-        supraRegistryId,
-        coinType
-      );
-    }
-    if (rule.includes('switchboard')) {
-      updateSwitchboardPrice(
-        type,
-        txBlock,
-        switchboardPackageId,
-        request,
-        switchboardAggregatorId,
-        switchboardRegistryId,
-        coinType
-      );
-    }
-  });
+  if (rules.includes('pyth')) {
+    updatePythPrice(
+      txBlock,
+      pythPackageId,
+      request,
+      pythStateId,
+      pythFeedObjectId,
+      pythRegistryId,
+      coinType
+    );
+  }
+  if (rules.includes('switchboard')) {
+    updateSwitchboardPrice(
+      txBlock,
+      switchboardPackageId,
+      request,
+      switchboardAggregatorId,
+      switchboardRegistryId,
+      coinType
+    );
+  }
+  if (rules.includes('supra')) {
+    updateSupraPrice(
+      txBlock,
+      supraPackageId,
+      request,
+      supraHolderId,
+      supraRegistryId,
+      coinType
+    );
+  }
 
   confirmPriceUpdateRequest(
     txBlock,
@@ -266,7 +243,6 @@ const confirmPriceUpdateRequest = (
  * @return TxBlock created by SuiKit.
  */
 const updateSupraPrice = (
-  type: xOracleCategoryType,
   txBlock: SuiKitTxBlock,
   packageId: string,
   request: TransactionArgument,
@@ -275,7 +251,7 @@ const updateSupraPrice = (
   coinType: string
 ) => {
   txBlock.moveCall(
-    `${packageId}::rule::set_price_as_${type}`,
+    `${packageId}::rule::set_price`,
     [request, holderId, registryId, SUI_CLOCK_OBJECT_ID],
     [coinType]
   );
@@ -294,7 +270,6 @@ const updateSupraPrice = (
  * @return TxBlock created by SuiKit.
  */
 const updateSwitchboardPrice = (
-  type: xOracleCategoryType,
   txBlock: SuiKitTxBlock,
   packageId: string,
   request: TransactionArgument,
@@ -303,7 +278,7 @@ const updateSwitchboardPrice = (
   coinType: string
 ) => {
   txBlock.moveCall(
-    `${packageId}::rule::set_price_as_${type}`,
+    `${packageId}::rule::set_price`,
     [request, aggregatorId, registryId, SUI_CLOCK_OBJECT_ID],
     [coinType]
   );
@@ -325,7 +300,6 @@ const updateSwitchboardPrice = (
  * @return TxBlock created by SuiKit.
  */
 const updatePythPrice = (
-  type: xOracleCategoryType,
   txBlock: SuiKitTxBlock,
   packageId: string,
   request: TransactionArgument,
@@ -335,7 +309,7 @@ const updatePythPrice = (
   coinType: string
 ) => {
   txBlock.moveCall(
-    `${packageId}::rule::set_price_as_${type}`,
+    `${packageId}::rule::set_price`,
     [request, stateId, feedObjectId, registryId, SUI_CLOCK_OBJECT_ID],
     [coinType]
   );