+++ src/models/scallopAddress.ts
@@ -1,9 +1,10 @@
 import { API_BASE_URL, USE_TEST_ADDRESS } from '../constants';
-import type { NetworkType } from '@scallop-io/sui-kit';
+import { SuiKit, type NetworkType } from '@scallop-io/sui-kit';
 import type {
   ScallopAddressParams,
   AddressesInterface,
   AddressStringPath,
+  ScallopAddressInstanceParams,
 } from '../types';
 import { ScallopCache } from './scallopCache';
 import { DEFAULT_CACHE_OPTIONS } from 'src/constants/cache';
@@ -33,7 +34,7 @@ const EMPTY_ADDRESSES: AddressesInterface = {
           },
         },
       },
-      apt: {
+      wapt: {
         id: '',
         metaData: '',
         treasury: '',
@@ -46,7 +47,7 @@ const EMPTY_ADDRESSES: AddressesInterface = {
           },
         },
       },
-      sol: {
+      wsol: {
         id: '',
         metaData: '',
         treasury: '',
@@ -59,7 +60,7 @@ const EMPTY_ADDRESSES: AddressesInterface = {
           },
         },
       },
-      btc: {
+      wbtc: {
         id: '',
         metaData: '',
         treasury: '',
@@ -72,7 +73,7 @@ const EMPTY_ADDRESSES: AddressesInterface = {
           },
         },
       },
-      eth: {
+      weth: {
         id: '',
         metaData: '',
         treasury: '',
@@ -85,7 +86,7 @@ const EMPTY_ADDRESSES: AddressesInterface = {
           },
         },
       },
-      usdc: {
+      wusdc: {
         id: '',
         metaData: '',
         treasury: '',
@@ -98,7 +99,7 @@ const EMPTY_ADDRESSES: AddressesInterface = {
           },
         },
       },
-      usdt: {
+      wusdt: {
         id: '',
         metaData: '',
         treasury: '',
@@ -237,7 +238,7 @@ const EMPTY_ADDRESSES: AddressesInterface = {
     adminCap: '',
     object: '',
     pools: {
-      seth: {
+      sweth: {
         id: '',
         rewardPoolId: '',
       },
@@ -245,11 +246,11 @@ const EMPTY_ADDRESSES: AddressesInterface = {
         id: '',
         rewardPoolId: '',
       },
-      susdc: {
+      swusdc: {
         id: '',
         rewardPoolId: '',
       },
-      susdt: {
+      swusdt: {
         id: '',
         rewardPoolId: '',
       },
@@ -324,15 +325,15 @@ const EMPTY_ADDRESSES: AddressesInterface = {
         coinType: '',
         treasury: '',
       },
-      susdc: {
+      swusdc: {
         coinType: '',
         treasury: '',
       },
-      susdt: {
+      swusdt: {
         coinType: '',
         treasury: '',
       },
-      seth: {
+      sweth: {
         coinType: '',
         treasury: '',
       },
@@ -370,11 +371,21 @@ export class ScallopAddress {
   private _network: NetworkType;
   private _currentAddresses?: AddressesInterface;
   private _addressesMap: Map<NetworkType, AddressesInterface>;
-  private _cache: ScallopCache;
+  public cache: ScallopCache;
 
-  public constructor(params: ScallopAddressParams, cache?: ScallopCache) {
+  public constructor(
+    params: ScallopAddressParams,
+    instance?: ScallopAddressInstanceParams
+  ) {
     const { id, auth, network } = params;
-    this._cache = cache ?? new ScallopCache(DEFAULT_CACHE_OPTIONS);
+    this.cache =
+      instance?.cache ??
+      new ScallopCache(
+        instance?.suiKit ?? new SuiKit({}),
+        undefined,
+        DEFAULT_CACHE_OPTIONS
+      );
+
     this._requestClient = axios.create({
       baseURL: API_BASE_URL,
       headers: {
@@ -573,7 +584,7 @@ export class ScallopAddress {
   public async read(id?: string) {
     const addressesId = id || this._id || undefined;
     if (addressesId !== undefined) {
-      const response = await this._cache.queryClient.fetchQuery({
+      const response = await this.cache.queryClient.fetchQuery({
         queryKey: ['api-getAddresses', addressesId],
         queryFn: async () => {
           return await this._requestClient.get(`/addresses/${addressesId}`, {