+++ src/models/scallopClient.ts
@@ -13,13 +13,15 @@ import { ScallopUtils } from './scallopUtils';
 import { ScallopBuilder } from './scallopBuilder';
 import { ScallopQuery } from './scallopQuery';
 import type { SuiTransactionBlockResponse } from '@mysten/sui/client';
-import type { TransactionObjectArgument } from '@mysten/sui/transactions';
+import type {
+  TransactionObjectArgument,
+  TransactionResult,
+} from '@mysten/sui/transactions';
 import { ScallopCache } from './scallopCache';
 import { requireSender } from 'src/utils';
 import type { SuiObjectArg } from '@scallop-io/sui-kit';
 import type {
   ScallopClientFnReturnType,
-  ScallopInstanceParams,
   ScallopClientParams,
   SupportPoolCoins,
   SupportCollateralCoins,
@@ -29,6 +31,8 @@ import type {
   SupportBorrowIncentiveCoins,
   ScallopTxBlock,
   SupportSCoin,
+  ScallopClientVeScaReturnType,
+  ScallopClientInstanceParams,
 } from '../types';
 
 /**
@@ -56,48 +60,46 @@ export class ScallopClient {
 
   public constructor(
     params: ScallopClientParams,
-    instance?: ScallopInstanceParams
+    instance?: ScallopClientInstanceParams
   ) {
     this.params = params;
-    this.suiKit = instance?.suiKit ?? new SuiKit(params);
-    this.cache =
-      instance?.cache ?? new ScallopCache(DEFAULT_CACHE_OPTIONS, this.suiKit);
-    this.address =
-      instance?.address ??
-      new ScallopAddress(
+    this.suiKit =
+      instance?.suiKit ?? instance?.builder?.suiKit ?? new SuiKit(params);
+    this.walletAddress = normalizeSuiAddress(
+      params?.walletAddress || this.suiKit.currentAddress()
+    );
+
+    if (instance?.builder) {
+      this.builder = instance.builder;
+      this.query = this.builder.query;
+      this.utils = this.query.utils;
+      this.address = this.utils.address;
+      this.cache = this.address.cache;
+    } else {
+      this.cache = new ScallopCache(
+        this.suiKit,
+        this.walletAddress,
+        DEFAULT_CACHE_OPTIONS
+      );
+      this.address = new ScallopAddress(
         {
           id: params?.addressesId || ADDRESSES_ID,
           network: params?.networkType,
         },
-        this.cache
+        {
+          cache: this.cache,
+        }
       );
-    this.query =
-      instance?.query ??
-      new ScallopQuery(params, {
-        suiKit: this.suiKit,
+      this.utils = new ScallopUtils(this.params, {
         address: this.address,
-        cache: this.cache,
       });
-    this.utils =
-      instance?.utils ??
-      new ScallopUtils(params, {
-        suiKit: this.suiKit,
-        address: this.address,
-        query: this.query,
-        cache: this.cache,
-      });
-    this.builder =
-      instance?.builder ??
-      new ScallopBuilder(params, {
-        suiKit: this.suiKit,
-        address: this.address,
-        query: this.query,
+      this.query = new ScallopQuery(this.params, {
         utils: this.utils,
-        cache: this.cache,
       });
-    this.walletAddress = normalizeSuiAddress(
-      params?.walletAddress || this.suiKit.currentAddress()
-    );
+      this.builder = new ScallopBuilder(this.params, {
+        query: this.query,
+      });
+    }
   }
 
   /**
@@ -817,7 +819,7 @@ export class ScallopClient {
       await this.utils.mergeSimilarCoins(
         txBlock,
         coin,
-        this.utils.parseCoinType(this.utils.parseCoinName(stakeCoinName)),
+        this.utils.parseCoinType(stakeCoinName),
         requireSender(txBlock)
       );
 
@@ -989,6 +991,7 @@ export class ScallopClient {
    * @returns Transaction response
    */
   public async migrateAllMarketCoin<S extends boolean>(
+    includeStakePool: boolean = true,
     sign: S = true as S
   ): Promise<ScallopClientFnReturnType<S>> {
     const txBlock = this.builder.createTxBlock();
@@ -1040,18 +1043,20 @@ export class ScallopClient {
           );
           sCoins.push(sCoin);
         }
-        // check for staked market coin in spool
-        if (SUPPORT_SPOOLS.includes(sCoinName as SupportStakeMarketCoins)) {
-          try {
-            const sCoin = await txBlock.unstakeQuick(
-              Number.MAX_SAFE_INTEGER,
-              sCoinName as SupportStakeMarketCoins
-            );
-            if (sCoin) {
-              sCoins.push(sCoin);
+        if (includeStakePool) {
+          // check for staked market coin in spool
+          if (SUPPORT_SPOOLS.includes(sCoinName as SupportStakeMarketCoins)) {
+            try {
+              const sCoin = await txBlock.unstakeQuick(
+                Number.MAX_SAFE_INTEGER,
+                sCoinName as SupportStakeMarketCoins
+              );
+              if (sCoin) {
+                sCoins.push(sCoin);
+              }
+            } catch (e: any) {
+              // ignore
             }
-          } catch (e: any) {
-            // ignore
           }
         }
 
@@ -1076,6 +1081,67 @@ export class ScallopClient {
       )) as ScallopClientFnReturnType<S>;
     } else {
       return txBlock.txBlock as ScallopClientFnReturnType<S>;
+    }
+  }
+
+  /* ==================== VeSCA ==================== */
+  /**
+   * Claim unlocked SCA from all veSCA accounts.
+   */
+  public async claimAllUnlockedSca(): Promise<SuiTransactionBlockResponse>;
+  public async claimAllUnlockedSca<S extends boolean>(
+    sign?: S
+  ): Promise<ScallopClientVeScaReturnType<S>>;
+  public async claimAllUnlockedSca<S extends boolean>(
+    sign: S = true as S
+  ): Promise<ScallopClientVeScaReturnType<S>> {
+    // get all veSca keys
+    const veScaKeys = (
+      (await this.query.getVeScas(this.walletAddress)) ?? []
+    ).map(({ keyObject }) => keyObject);
+    if (veScaKeys.length === 0) {
+      throw new Error('No veSCA found in the wallet');
+    }
+
+    const scaCoins: TransactionResult[] = [];
+    const tx = this.builder.createTxBlock();
+    tx.setSender(this.walletAddress);
+
+    await Promise.all(
+      veScaKeys.map(async (key) => {
+        try {
+          const scaCoin = await tx.redeemScaQuick(key, false);
+          if (!scaCoin) return;
+          scaCoins.push(scaCoin);
+        } catch (e) {
+          // ignore
+        }
+      })
+    );
+
+    if (scaCoins.length === 0) {
+      throw new Error('No unlocked SCA found in the veSCA accounts');
+    }
+
+    if (scaCoins.length > 1) {
+      tx.mergeCoins(scaCoins[0], scaCoins.slice(1));
+    }
+    await this.utils.mergeSimilarCoins(
+      tx,
+      scaCoins[0],
+      'sca',
+      this.walletAddress
+    );
+
+    if (sign) {
+      return (await this.suiKit.signAndSendTxn(
+        tx
+      )) as ScallopClientVeScaReturnType<S>;
+    } else {
+      return {
+        tx: tx.txBlock,
+        scaCoin: scaCoins[0],
+      } as ScallopClientVeScaReturnType<S>;
     }
   }