+++ test/query.spec.ts
@@ -1,6 +1,5 @@
 import { describe, it, expect } from 'vitest';
 import { isValidSuiAddress } from '@scallop-io/sui-kit';
-import { getVescaKeys } from 'src/queries';
 import { z as zod } from 'zod';
 import { scallopSDK } from './scallopSdk';
 
@@ -20,7 +19,7 @@ describe('Test Query Scallop Contract On Chain Data', async () => {
   });
 
   it('Should get market pools data', async () => {
-    const marketPools = await scallopQuery.getMarketPools(['sui', 'usdc']);
+    const marketPools = await scallopQuery.getMarketPools(['sui', 'wusdc']);
 
     if (ENABLE_LOG) {
       console.info('Market pool:');
@@ -42,7 +41,7 @@ describe('Test Query Scallop Contract On Chain Data', async () => {
   it('Should get market collaterals data', async () => {
     const marketCollaterals = await scallopQuery.getMarketCollaterals([
       'sui',
-      'usdc',
+      'wusdc',
     ]);
 
     if (ENABLE_LOG) {
@@ -121,7 +120,7 @@ describe('Test Query Scallop Contract On Chain Data', async () => {
   });
 
   it('Should get pyth price data', async () => {
-    const usdcPrice = await scallopQuery.getPriceFromPyth('usdc');
+    const usdcPrice = await scallopQuery.getPriceFromPyth('wusdc');
 
     if (ENABLE_LOG) {
       console.info('Usdc price:', usdcPrice);
@@ -251,7 +250,7 @@ describe('Test Portfolio Query', async () => {
   console.info('Your wallet:', scallopQuery.suiKit.currentAddress());
 
   it('Should get user lendings data', async () => {
-    const lendings = await scallopQuery.getLendings(['sui', 'usdc']);
+    const lendings = await scallopQuery.getLendings(['sui', 'wusdc']);
 
     if (ENABLE_LOG) {
       console.info('User lendings:', lendings);
@@ -310,16 +309,14 @@ describe('Test VeSca Query', async () => {
   const sender = scallopQuery.suiKit.currentAddress();
   console.info(`Your Wallet: ${sender}`);
 
-  const veScaKeys = await getVescaKeys(scallopQuery, sender);
   let obligationId: string | undefined;
-  if (veScaKeys.length === 0)
-    throw new Error(`No VeSca keys found in ${sender}`);
 
+  const VE_SCA_KEY =
+    '0xad50994e23ae4268fc081f477d0bdc3f1b92c7049c9038dedec5bac725273d18' as const;
   // make sure you test with an account that has binded obligationId to a veScaKey
-  it(`Should get binded obligationId of veScaKey ${veScaKeys[0].objectId}`, async () => {
-    const bindedObligationId = await scallopQuery.getBindedObligationId(
-      veScaKeys[0].objectId
-    );
+  it(`Should get binded obligationId of veScaKey ${VE_SCA_KEY}`, async () => {
+    const bindedObligationId =
+      await scallopQuery.getBindedObligationId(VE_SCA_KEY);
 
     if (ENABLE_LOG) {
       console.info('Binded Obligation Id:', bindedObligationId);
@@ -327,7 +324,7 @@ describe('Test VeSca Query', async () => {
 
     if (!bindedObligationId)
       throw new Error(
-        `No binded obligationId found for veScaKey ${veScaKeys[0]}`
+        `No binded obligationId found for veScaKey ${VE_SCA_KEY}`
       );
     obligationId = bindedObligationId;
     expect(!!bindedObligationId).toBe(true);
@@ -362,6 +359,24 @@ describe('Test VeSca Query', async () => {
       averageLockingPeriodUnit: zod.string(),
     });
     expect(treasuryInfoSchema.safeParse(totalVeScaTreasury).success).toBe(true);
+  });
+
+  it(`Should get veSCA`, async () => {
+    const veSca = await scallopQuery.getVeSca(VE_SCA_KEY);
+    if (ENABLE_LOG) {
+      console.info('VeSca:', veSca);
+    }
+    expect(!!veSca).toBe(true);
+
+    const veScaSchema = zod.object({
+      id: zod.string(),
+      keyId: zod.string(),
+      lockedScaAmount: zod.string(),
+      lockedScaCoin: zod.number(),
+      currentVeScaBalance: zod.number(),
+      unlockAt: zod.number(),
+    });
+    expect(veScaSchema.safeParse(veSca).success).toBe(true);
   });
 
   it(`Should get veSCAs`, async () => {
@@ -434,5 +449,31 @@ describe('Test sCoin Query', async () => {
 
     expect(typeof totalSupply).toBe('number');
     expect(totalSupply >= 0).toBe(true);
+  });
+
+  it('Should get swap rate between sCoin assets', async () => {
+    const fromSCoin = 'swusdc';
+    const toSCoin = 'ssui';
+
+    const getSCoinSwapRate = await scallopQuery.getSCoinSwapRate(
+      fromSCoin,
+      toSCoin
+    );
+    console.log('getSCoinSwapRate', getSCoinSwapRate);
+    expect(typeof getSCoinSwapRate).toBe('number');
+    expect(getSCoinSwapRate > 0).toBe(true);
+  });
+});
+
+describe('Test Supply Limit Query', async () => {
+  const scallopQuery = await scallopSDK.createScallopQuery();
+  const sender = scallopQuery.suiKit.currentAddress();
+  console.info(`Your Wallet: ${sender}`);
+
+  it('Should get supply limit of a pool', async () => {
+    const poolName = 'sui';
+    const supplyLimit = await scallopQuery.getPoolSupplyLimit(poolName);
+    expect(typeof supplyLimit).toBe('string');
+    expect(+supplyLimit! > 0).toBe(true);
   });
 });